! Fix save shipmentDetail failed

See merge request huydp_Irtech/eport!367

+++ eport-logistic/eport-logistic-portal/src/main/java/vn/com/irtech/eport/logistic/controller/LogisticSendContEmptyController.java
@@ -158,9 +158,14 @@ public class LogisticSendContEmptyController extends LogisticBaseController {
 	public AjaxResult saveShipmentDetail(@RequestBody List<ShipmentDetail> shipmentDetails) {
 		if (shipmentDetails != null) {
 			LogisticAccount user = getUser();
-			int index = 0;
+			List<String> contReservedList = shipmentDetailService.checkContainerReserved(shipmentDetails.get(0).getProcessStatus());
+			if (contReservedList.size() > 0) {
+				AjaxResult ajaxResult = AjaxResult.error();
+				ajaxResult.put("conts", contReservedList);
+				return ajaxResult;
+			}
 			for (ShipmentDetail shipmentDetail : shipmentDetails) {
-				index++;
+				shipmentDetail.setProcessStatus(null);
 				if (shipmentDetail.getId() != null) {
 					if (shipmentDetail.getContainerNo() == null || shipmentDetail.getContainerNo().equals("")) {
 						shipmentDetailService.deleteShipmentDetailById(shipmentDetail.getId());
@@ -175,13 +180,9 @@ public class LogisticSendContEmptyController extends LogisticBaseController {
 					shipmentDetail.setLogisticGroupId(user.getGroupId());
 					shipmentDetail.setCreateBy(user.getFullName());
 					shipmentDetail.setCreateTime(new Date());
-					shipmentDetail.setRegisterNo(shipmentDetail.getShipmentId().toString()+index);
 					shipmentDetail.setStatus(1);
 					shipmentDetail.setPaymentStatus("N");
 					shipmentDetail.setProcessStatus("N");
-					shipmentDetail.setWgt(1l);
-					shipmentDetail.setVslNm("vslNm");
-					shipmentDetail.setVoyNo("voyNo");
 					shipmentDetail.setFe("E");
 					if(shipmentDetail.getLoadingPort() == null) {
 						shipmentDetail.setLoadingPort(" ");
